infrastructure/azure-functions/requirements.txt
# Azure Functions
azure-functions>=1.17.0
azure-storage-blob>=12.19.0

# OpenAI
openai>=1.12.0

# Data processing
pydantic>=2.5.0
python-dotenv>=1.0.0

# Crawler dependencies (if running analysis in function)
# Note: These may need to be adjusted based on your actual analysis package
aiohttp>=3.9.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
